+++ lib/react_on_rails/test_helper/ensure_assets_compiled.rb
@@ -29,7 +29,7 @@ module ReactOnRails
         # Be sure we don't do this again.
         self.class.has_been_run = true
 
-        ReactOnRails::LocalesToJs.new
+        ReactOnRails::Locales::ToJs.new
 
         stale_gen_files = webpack_assets_status_checker.stale_generated_webpack_files